Moved all the server container logic to the container behavior implementation.
